fix: `service_endpoints` has been renamed to `kms_endpoint` and only allows a value of `public` or `private` - `kms_key_crn` has been renamed to `kms_key_id` and now requires the Key Proect Key ID not be passed in, and not the CRN as before (#51)
BREAKING CHANGE: If you provisioned an instance in a previous release and passed a Key Protect CRN value into the `kms_key_id` variable, your instance was probably in a unusable state. Although terraform will allow you to attempt to upgrade by now passing in the KMS Key ID, the upgrade will fail with the error: `Error updating resource instance: Please contact the Service Provider for this error. [400, Bad Request] Instance value 'plan_id' is missing`.
The only way to workaround this is to do a terraform destroy on the old instance, and recreate it using this version of the module or later. The data engine team are going to add extra validation to ensure going forward instance creation will fail if an incorrect value is passed for the KMS Key.